Extract get mailbox by student use case from mailbox repo
