+++ app/src/main/java/io/github/wulkanowy/data/repositories/MailboxRepository.kt
@@ -4,6 +4,7 @@ import io.github.wulkanowy.data.db.dao.MailboxDao
 import io.github.wulkanowy.data.db.entities.Mailbox
 import io.github.wulkanowy.data.db.entities.Student
 import io.github.wulkanowy.data.mappers.mapToEntities
+import io.github.wulkanowy.domain.messages.GetMailboxByStudentUseCase
 import io.github.wulkanowy.sdk.Sdk
 import io.github.wulkanowy.utils.AutoRefreshHelper
 import io.github.wulkanowy.utils.getRefreshKey
@@ -17,6 +18,7 @@ class MailboxRepository @Inject constructor(
     private val mailboxDao: MailboxDao,
     private val sdk: Sdk,
     private val refreshHelper: AutoRefreshHelper,
+    private val getMailboxByStudentUseCase: GetMailboxByStudentUseCase,
 ) {
     private val cacheKey = "mailboxes"
 
@@ -33,11 +35,11 @@ class MailboxRepository @Inject constructor(
     suspend fun getMailbox(student: Student): Mailbox {
         val isExpired = refreshHelper.shouldBeRefreshed(getRefreshKey(cacheKey, student))
         val mailboxes = mailboxDao.loadAll(student.userLoginId)
-        val mailbox = mailboxes.filterByStudent(student)
+        val mailbox = getMailboxByStudentUseCase(student)
 
         return if (isExpired || mailbox == null) {
             refreshMailboxes(student)
-            val newMailbox = mailboxDao.loadAll(student.userLoginId).filterByStudent(student)
+            val newMailbox = getMailboxByStudentUseCase(student)
 
             requireNotNull(newMailbox) {
                 "Mailbox for ${student.userName} - ${student.studentName} not found! Saved mailboxes: $mailboxes"
@@ -45,41 +47,5 @@ class MailboxRepository @Inject constructor(
 
             newMailbox
         } else mailbox
-    }
-
-    private fun List<Mailbox>.filterByStudent(student: Student): Mailbox? {
-        val normalizedStudentName = student.studentName.normalizeStudentName()
-
-        return find {
-            it.studentName.normalizeStudentName() == normalizedStudentName
-        } ?: singleOrNull {
-            it.studentName.getFirstAndLastPart() == normalizedStudentName.getFirstAndLastPart()
-        } ?: singleOrNull {
-            it.studentName.getUnauthorizedVersion() == normalizedStudentName
-        }
-    }
-
-    private fun String.normalizeStudentName(): String {
-        return trim().split(" ")
-            .filter { it.isNotBlank() }
-            .joinToString(" ") { part ->
-                part.lowercase().replaceFirstChar { it.uppercase() }
-            }
-    }
-
-    private fun String.getFirstAndLastPart(): String {
-        val parts = normalizeStudentName().split(" ")
-
-        val endParts = parts.filterIndexed { i, _ ->
-            i == 0 || parts.size - 1 == i
-        }
-        return endParts.joinToString(" ")
-    }
-
-    private fun String.getUnauthorizedVersion(): String {
-        return normalizeStudentName().split(" ")
-            .joinToString(" ") {
-                it.first() + "*".repeat(it.length - 1)
-            }
     }
 }
